Split up DashboardItem into multiple files
